requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
requests==2.32.3
Jinja2==3.1.4
itsdangerous==2.2.0